pyproject.toml
[tool.poetry]
name = "depenses"
version = "0.1.0"
description = "app for manage personal or group spendings"
authors = ["Mikita Chabatarovich <mikita.chabatarovich@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.7"
django-money = "^1.1"
djangorestframework = "^3.11.0"
djoser = "^2.0.3"

[tool.poetry.dev-dependencies]
tox = "^3.15.2"
black = "^19.10b0"
pytest-django = "^3.9.0"
invoke = "^1.4.1"

[tool.black]
line-length = 120
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

